# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Xihe API
  version: 1.0.0
  description: Canonical HTTP contract for Xihe public and service APIs.
servers:
  - url: /
paths:
  /api/v1/auth/register:
    post:
      operationId: register
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RegisterRequest' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }, '400': { $ref: '#/components/responses/Problem' } }
  /api/v1/auth/login:
    post:
      operationId: login
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } } }
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }, '401': { $ref: '#/components/responses/Problem' } }
  /api/v1/auth/refresh:
    post:
      operationId: refresh
      responses: { '200': { description: OK }, '401': { $ref: '#/components/responses/Problem' } }
  /api/v1/auth/me:
    get:
      operationId: currentUser
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: OK }, '401': { $ref: '#/components/responses/Problem' } }
...
  /internal/v1/agent/models:
    get: { operationId: agentModels, security: [{ bearerAuth: [] }], responses: { '200': { description: Models } } }
  /internal/v1/agent/embedding-models:
    get: { operationId: agentEmbeddingModels, security: [{ bearerAuth: [] }], responses: { '200': { description: Embedding models } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } }
  schemas:
    RegisterRequest: { type: object, required: [email, password, name], properties: { email: { type: string }, password: { type: string }, name: { type: string } } }
    LoginRequest: { type: object, required: [email, password], properties: { email: { type: string }, password: { type: string } } }
    AuthResponse: { type: object, properties: { accessToken: { type: string }, refreshToken: { type: string }, user: { type: object } } }
...
    ProblemDetails:
      type: object
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        requestId: { type: string }